requirements.txt
flask 
jsonpickle 
opencv-python 
pillow
matplotlib
gunicorn==19.9.0
